Move preserveMegaGroups to heapAlloc
