+++ asterius/rts/rts.mblockalloc.mjs
@@ -78,25 +78,4 @@ export class MBlockAlloc {
       this.freeList.push(bd);
     }
   }
-
-  preserveMegaGroups(bds) {
-    this.freeList = [];
-    const sorted_bds = Array.from(bds).sort((bd0, bd1) => bd0 - bd1);
-    sorted_bds.push(Memory.tagData(rtsConstants.mblock_size * this.capacity) + rtsConstants.offset_first_bdescr);
-    this.freeSegment(
-        Memory.tagData(rtsConstants.mblock_size * this.staticMBlocks),
-        sorted_bds[0] - rtsConstants.offset_first_bdescr);
-    for (let i = 0; i < (sorted_bds.length-1); ++i) {
-      const l_start = Number(
-          this.memory.i64Load(sorted_bds[i] + rtsConstants.offset_bdescr_start)),
-      l_blocks =
-          this.memory.i32Load(sorted_bds[i] + rtsConstants.offset_bdescr_blocks),
-      l_end = l_start + (rtsConstants.block_size * l_blocks),
-      r = sorted_bds[i + 1] - rtsConstants.offset_first_bdescr;
-      this.freeSegment(l_end, r);
-    }
-    this.freeList.sort(
-        (bd0, bd1) => this.memory.i32Load(bd0 + rtsConstants.offset_bdescr_blocks) -
-                  this.memory.i32Load(bd1 + rtsConstants.offset_bdescr_blocks));
-  }
 }
